Move tests from src to tests folder
