+++ build.config.ts
@@ -2,5 +2,13 @@ import {defineBuildConfig} from "unbuild"
 
 export default defineBuildConfig({
   declaration: false,
-  outDir: "lib"
+  outDir: "lib",
+  entries: [
+    {
+      input: "src",
+      builder: "mkdist",
+      format: "cjs",
+      ext: "js"
+    }
+  ]
 })
