+++ src/configs/typescript.ts
@@ -2,7 +2,7 @@ import type {ESLint} from "eslint"
 
 export default {
   extends: [
-    "./base.cjs", // TODO: Add resolveConfigBase or emit cjs only
+    "./base.cjs",
     "plugin:@typescript-eslint/eslint-recommended"
   ],
   parser: "@typescript-eslint/parser",
