+++ tests/__macro__/withAssertRules.ts
