+++ tests/react.test.ts
@@ -3,10 +3,10 @@ import {fileURLToPath} from "node:url"
 import js from "dedent"
 import test from "ava"
 
-import {withAssertRules} from "../__macro__/withAssertRules.js"
+import {withAssertRules} from "./__macro__/withAssertRules.js"
 
 const configPath = fileURLToPath(
-  new URL("../../lib/configs/react.cjs", import.meta.url)
+  new URL("../lib/configs/react.cjs", import.meta.url)
 )
 
 test("Lints basic code example", withAssertRules, {
