+++ tests/rules/base/arrow-parens.test.ts
