+++ tests/rules/base/camelcase.test.ts
