+++ tests/rules/base/no-plusplus.test.ts
