+++ tests/rules/base/no-semi.test.ts
