+++ tests/rules/base/operator-linebreak.test.ts
