+++ tests/rules/base/prefer-const.test.ts
