+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  "include": ["**/*.ts"],
   "compilerOptions": {
     "target": "esnext",
     "jsx": "preserve",
